pyproject.toml
[tool.poetry]
name = "team 3 project"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Yaroslav Kubitskyi <fenixm22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
uvicorn = "^0.31.0"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.12"
bcrypt = "^4.2.0"
qrcode = "^8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
